+++ CHANGELOG.md
@@ -8,10 +8,10 @@
 ### Bug
 
 - Deep watch on chart data sets [#885](https://github.com/EthVM/EthVM/pull/885)
+- Fix mobile loading state padding for address header page [#891](https://github.com/EthVM/EthVM/pull/891)
 - Change 24 price change to 24 price change percentage [#887](https://github.com/EthVM/EthVM/pull/887)
 - Sentry fix #1950193750 and set contract correctly in CoinData [#895](https://github.com/EthVM/EthVM/pull/895)
 - Catch v-img errors on load [#881](https://github.com/EthVM/EthVM/pull/881) 
-
 
 ===================================================================================